+++ platforms/all/fbpic/stage/libe_opt/ensemble_runner.py
@@ -7,7 +7,7 @@ from libensemble import logger
 from libensemble.executors.mpi_executor import MPIExecutor
 from libensemble.tools import save_libE_output, add_unique_random_streams
 
-from utils import (
+from libe_opt.utils import (
     create_alloc_specs, create_gen_specs, create_libe_specs, create_sim_specs)
