+++ platforms/all/fbpic/stage/template_simulation_script.py
