--- platforms/swing/fbpic/2n_8t_8gpu.json
@@ -1,17 +1,0 @@
-{
-    "calling": {
-        "gen_type": "'aposmm'",
-        "template": "run_libe_fbpic.py",
-        "run_async": true,
-        "sim_max": 20
-    },
-
-    "submit": {
-        "num_nodes": 2,
-        "num_tasks": 8,
-        "num_gpus": 8,
-        "num_workers": 9,
-        "omp_num_threads": 1,
-        "template": "base_template"
-    }
-}
